Added vcan1 to send ignition to vcandump
