+++ App/Emulator/CMakeLists.txt
@@ -7,11 +7,12 @@ project(CAN_Reader VERSION 1.0)
 add_executable(vcanReader
             main.cpp
             src/emulator.cpp
+            src/CAN_reader.cpp
+            src/decoding.cpp
 )
 
 
 target_link_libraries(vcanReader PUBLIC socketCanlib)
 target_include_directories(vcanReader PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
 
-target_link_libraries(inputHandler PUBLIC ${CURSES_LIBRARIES})
