--- App/Emulator/include/efuntions.h
@@ -1,6 +1,0 @@
-#ifndef EFUNTIONS_H
-#define EFUNTIONS_H
-
-void Reader();
-
-#endif  //EFUNTIONS_H
