+++ App/Emulator/include/reader.h
@@ -1,0 +1,11 @@
+#ifndef EFUNTIONS_H
+#define EFUNTIONS_H
+
+#include "socketcan_cpp.h"
+
+void Reader();
+//testing for sending to dashboard
+void SendToDashboard(int);
+
+
+#endif  //EFUNTIONS_H
