+++ App/Emulator/main.cpp
@@ -1,10 +1,9 @@
 #include <iostream>
-#include "efuntions.h"
+#include "reader.h"
 
 //#include "socketcan_cpp.h"
 // https://github.com/siposcsaba89/socketcan-cpp
 
-using namespace std;
 
 int main() {
